+++ operator/src/lib.rs
@@ -35,7 +35,7 @@ pub mod metrics_scraper;
 /// Kubernetes resource builders.
 pub mod resources;
 
-pub use resources::trust_bundle::sha256_fingerprint;
+pub use resources::{tls_backend::init_process_crypto, trust_bundle::sha256_fingerprint};
 /// Provider gateway address self-discovery.
 pub mod gateway;
 /// SWIM membership data model and status summarization.
